Updated render

+++ app/src/main/java/com/eszdman/photoncamera/Render/Parameters.java
@@ -8,7 +8,7 @@ import android.hardware.camera2.params.BlackLevelPattern;
 public class Parameters {
     public byte cfaPattern;
     public Point rawSize;
-    public byte[] blacklevel = new byte[4];
+    public float[] blacklevel = new float[4];
     public float[] whitepoint = new float[3];
     public short whitelevel = 1023;
     public String path;
@@ -18,7 +18,7 @@ public class Parameters {
         for(int i =0; i<4; i++) blacklevel[i] = 64;
         int[] blarr = new int[4];
         BlackLevelPattern level = characteristics.get(CameraCharacteristics.SENSOR_BLACK_LEVEL_PATTERN);
-        if(level != null) for(int i =0; i<4;i++) blacklevel[i] = (byte)blarr[i];
+        if(level != null) for(int i =0; i<4;i++) blacklevel[i] = blarr[i];
         Object ptr = characteristics.get(CameraCharacteristics.SENSOR_INFO_COLOR_FILTER_ARRANGEMENT);
         if(ptr !=null) cfaPattern = (byte)(int)ptr;
         Object wlevel = characteristics.get(CameraCharacteristics.SENSOR_INFO_WHITE_LEVEL);
